+++ src/.vuepress/data/index.js
@@ -1,6 +1,8 @@
 const dataFiles = require.context(".", false, /\.json$/);
+const presentationDataFiles = require.context("./presentations", true, /\.json$/);
 
 const data = {};
+const presentationData = {};
 
 dataFiles.keys().forEach((fileName) => {
   const fileData = dataFiles(fileName);
@@ -8,4 +10,20 @@ dataFiles.keys().forEach((fileName) => {
   data[year] = fileData.default || fileData;
 });
 
+presentationDataFiles.keys().forEach((fileName) => {
+  let fileData = presentationDataFiles(fileName);
+  fileName = fileName.replace(/\.\//, '').replace(/\.json/, '')
+  const [theme, year, title] = fileName.split("/", 3)
+  if (!presentationData[theme]) {
+    presentationData[theme] = {}
+  }
+  if (!presentationData[theme][year]) {
+    presentationData[theme][year] = []
+  }
+  fileData = fileData.default || fileData;
+  fileData = {...fileData, title}
+  presentationData[theme][year].push(fileData.default || fileData)
+});
+
 export default data;
+export { data, presentationData};