+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,39 +7,65 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 1.0.1
+
+- Add section to guide on how to get started streaming an event
+
+## 1.0.0
+
+- Update README with new design and setup information
+- Update to VuePress 1.2.0
+- Upgrade theme to new default theme in VuePress 1.2.0
+
 ## 0.1.0-beta - 2019.04.22
+
 ### Added
+
 - Easier mechanism for managing CFPs
 
 ## 0.5.0-alpha - 2019.04.21
+
 ### Added
+
 - New Event issue template
+
 ### Changed
+
 - Made file naming consistent for issue templates
 
 ## 0.4.0-alpha - 2019.02.26
+
 ### Added
+
 - Call for Proposals section
 - Add GitHub issue templates
 
 ## 0.3.0-alpha - 2019.02.23
+
 ### Added
+
 - Netlify deployment badge on README
 - Site version badge on README
 - Dedicated conferences page
 
 ## 0.2.0-alpha - 2019.02.17
+
 ### Added
+
 - Repo link in navigation for easy access
 - Edit link on the bottom of page in order to make it easy to contribute to the project
 - New workshop section for in-person training
 
 ## 0.1.1-alpha - 2019.02.07
+
 ### Added
+
 - Migrate data from VueMeetups.org
 
 ## 0.1.0-alpha - 2019.01.29
+
 ### Added
+
 - Setup initial VuePress installation
 - Add in baseline theme for future customization
 - Google Analytics plugin with tracking ID